refactor: move collaboration components into features with stubs
